allow webhooks to be added and move the alertmanager endpoint
