+++ src/iris/api.py
@@ -35,6 +35,7 @@ from iris.sender import auditlog
 from iris.sender.quota import (get_application_quotas_query, insert_application_quota_query,
                                required_quota_keys, quota_int_keys)
 
+from iris.custom_import import import_custom_module
 
 from .constants import (
     XFRAME, XCONTENTTYPEOPTIONS, XXSSPROTECTION
@@ -669,20 +670,13 @@ class AuthMiddleware(object):
 
         # Proceed with authenticating this route as a third party application
         try:
-            # Ignore HMAC requirements for Alertmanager, which doesn't support it
-            if len(req.env['PATH_INFO'].split("/")) > 2 and req.env['PATH_INFO'].split("/")[2] == 'alertmanager' and method == 'POST':
+            # Ignore HMAC requirements for custom webhooks
+            if len(req.env['PATH_INFO'].split("/")) > 2 and req.env['PATH_INFO'].split("/")[2] == 'webhooks'
                 qs = parse_qs(req.env['QUERY_STRING'])
 
                 app = qs['application'][0]
                 if not app:
                     raise HTTPBadRequest('Missing application keyvalue in query string')
-
-                # extract the target plan from the body, groupLabels, iris_plan label
-                alert_params = ujson.loads(req.context['body'])
-                if 'groupLabels' not in alert_params or 'iris_plan' not in alert_params['groupLabels']:
-                    raise HTTPBadRequest('Missing iris_plan label in groupLabels, in body')
-
-                req.context['plan'] = alert_params['groupLabels']['iris_plan']
             else:
                 app, client_digest = req.get_header('AUTHORIZATION', '')[5:].split(':', 1)
 
@@ -701,8 +695,8 @@ class AuthMiddleware(object):
         if resource.allow_read_no_auth and method == 'GET':
             return
 
-        # Ignore HMAC requirements for Alertmanager, which doesn't support it
-        if len(req.env['PATH_INFO'].split("/")) > 2 and req.env['PATH_INFO'].split("/")[2] == 'alertmanager' and method == 'POST':
+        # Ignore HMAC requirements for custom webhooks
+        if len(req.env['PATH_INFO'].split("/")) > 2 and req.env['PATH_INFO'].split("/")[2] == 'webhook'
             qs = parse_qs(req.env['QUERY_STRING'])
 
             app_name = qs['application'][0]
@@ -721,13 +715,6 @@ class AuthMiddleware(object):
             if not api_key == str(app['key']):
                 logger.warn('Application key invalid')
                 raise HTTPUnauthorized('Authentication failure', '', [])
-
-            # extract the target plan from the body, groupLabels, iris_plan label
-            alert_params = ujson.loads(req.context['body'])
-            if 'groupLabels' not in alert_params or 'iris_plan' not in alert_params['groupLabels']:
-                raise HTTPBadRequest('Missing iris_plan label in groupLabels, in body')
-
-            req.context['plan'] = alert_params['groupLabels']['iris_plan']
             return
 
         # If we're authenticated using beaker, don't validate app as if this is an
@@ -1277,112 +1264,6 @@ class Plans(object):
         resp.status = HTTP_201
         resp.body = ujson.dumps(plan_id)
         resp.set_header('Location', '/plans/%s' % plan_id)
-
-
-class Alertmanager(object):
-    allow_read_no_auth = False
-
-    def on_get(self, req, resp):
-        raise HTTPNotFound()
-
-    def on_post(self, req, resp):
-        '''
-        This endpoint is compatible with the webhook post from Alertmanager.
-        Simply configure alertmanager with a receiver pointing to iris, like
-        so:
-
-        receivers:
-        - name: 'iris-team1'
-          webhook_configs:
-            - url: http://iris:16649/v0/alertmanager?application=test-app&key=sdffdssdf
-
-        Where application points to an application and key it's key, in Iris.
-
-        For every POST from alertmanager, a new incident will be created.
-        '''
-        alert_params = ujson.loads(req.context['body'])
-        if not all(k in alert_params for k in ("version", "status", "alerts")):
-            raise HTTPBadRequest('missing version, status and/or alert attributes')
-
-        with db.guarded_session() as session:
-            plan = req.context['plan']
-            plan_id = session.execute('SELECT `plan_id` FROM `plan_active` WHERE `name` = :plan',
-                                      {'plan': plan}).scalar()
-            if not plan_id:
-                raise HTTPNotFound()
-
-            app = req.context['app']
-
-            # create new context
-            context = {}
-            str_fields = ['status', 'groupKey', 'version', 'receiver', 'externalURL']
-
-            for field in str_fields:
-                context[field] = alert_params[field]
-
-            cnt = 0
-            for alert in alert_params['alerts']:
-                cnt = cnt + 1
-                field_prefix = "alert_" + str(cnt) + "_"
-                context[field_prefix + "status"] = alert['status']
-                context[field_prefix + "endsAt"] = alert['endsAt']
-                context[field_prefix + "generatorURL"] = alert['generatorURL']
-                context[field_prefix + "startsAt"] = alert['startsAt']
-
-                labels = "; ".join([k + ": " + v for k, v in alert['labels'].items()])
-                context[field_prefix + "labels"] = labels
-
-                annotations = "; ".join([k + ": " + v for k, v in alert['annotations'].items()])
-                context[field_prefix + "annotations"] = annotations
-
-            for k, val in alert_params['groupLabels'].iteritems():
-                context['groupLabel_' + k] = val
-
-            for k, val in alert_params['commonLabels'].iteritems():
-                context['commonLabels_' + k] = val
-
-            for k, val in alert_params['commonAnnotations'].iteritems():
-                context['commonAnnotations_' + k] = val
-
-            context_json_str = ujson.dumps(context)
-            if len(context_json_str) > 65535:
-                logger.warn('POST to alertmanager exceeded acceptable size')
-                raise HTTPBadRequest('Context too long')
-
-            app_template_count = session.execute('''
-                SELECT EXISTS (
-                  SELECT 1 FROM
-                  `plan_notification`
-                  JOIN `template` ON `template`.`name` = `plan_notification`.`template`
-                  JOIN `template_content` ON `template_content`.`template_id` = `template`.`id`
-                  WHERE `plan_notification`.`plan_id` = :plan_id
-                  AND `template_content`.`application_id` = :app_id
-                )
-            ''', {'app_id': app['id'], 'plan_id': plan_id}).scalar()
-
-            if not app_template_count:
-                raise HTTPBadRequest('No plan template actions exist for this app')
-
-            data = {
-                'plan_id': plan_id,
-                'created': datetime.datetime.utcnow(),
-                'application_id': app['id'],
-                'context': context_json_str,
-                'current_step': 0,
-                'active': True,
-            }
-
-            incident_id = session.execute(
-                '''INSERT INTO `incident` (`plan_id`, `created`, `context`,
-                     `current_step`, `active`, `application_id`)
-                VALUES (:plan_id, :created, :context, 0, :active, :application_id)''',
-                data).lastrowid
-
-            session.commit()
-            session.close()
-        resp.status = HTTP_201
-        resp.set_header('Location', '/incidents/%s' % incident_id)
-        resp.body = ujson.dumps(incident_id)
 
 
 class Incidents(object):
@@ -3987,7 +3868,7 @@ def json_error_serializer(req, resp, exception):
     resp.content_type = 'application/json'
 
 
-def construct_falcon_api(debug, healthcheck_path, allowed_origins, iris_sender_app, zk_hosts, default_sender_addr):
+def construct_falcon_api(debug, healthcheck_path, allowed_origins, iris_sender_app, zk_hosts, default_sender_addr, config):
     cors = CORS(allow_origins_list=allowed_origins)
     api = API(middleware=[
         ReqBodyMiddleware(),
@@ -4004,8 +3885,6 @@ def construct_falcon_api(debug, healthcheck_path, allowed_origins, iris_sender_a
 
     api.add_route('/v0/incidents/{incident_id}', Incident())
     api.add_route('/v0/incidents', Incidents())
-
-    api.add_route('/v0/alertmanager', Alertmanager())
 
     api.add_route('/v0/messages/{message_id}', Message())
     api.add_route('/v0/messages/{message_id}/auditlog', MessageAuditLog())
@@ -4051,7 +3930,17 @@ def construct_falcon_api(debug, healthcheck_path, allowed_origins, iris_sender_a
     api.add_route('/v0/stats', Stats())
 
     api.add_route('/healthcheck', Healthcheck(healthcheck_path))
+
+    init_webhooks(config, api)
+
     return api
+
+
+def init_webhooks(config, api):
+    webhooks = config.get('webhooks', [])
+    for webhook in webhooks:
+        webhook_class = import_custom_module('iris.webhooks', webhook)
+        api.add_route('/v0/webhooks/' + webhook, webhook_class())
 
 
 def get_api(config):
@@ -4073,7 +3962,7 @@ def get_api(config):
 
     # all notifications go through master sender for now
     app = construct_falcon_api(
-        debug, healthcheck_path, allowed_origins, iris_sender_app, zk_hosts, default_master_sender_addr)
+        debug, healthcheck_path, allowed_origins, iris_sender_app, zk_hosts, default_master_sender_addr, config)
 
     # Need to call this after all routes have been created
     app = ui.init(config, app)
